Initial Commit
Now that this project is far enough in development, it is time to back
it up with good version control. This is a project to simplify basic
image manipulation. It is something I want to eventually use as a way to
play with inputs to neural networks. It also implements some cool
algorithms I found online for image processing and wanted to try as well
as some that I came up with myself. One of the other goals of this
project is to be able to count the 'blobs' in an image by manipulating
the colors and contrast via image processing.