Layed down graphics foundation for algorithm refactor.
The algorithm refactor is going to include getting rid of the current
iteration count system in favor of a bidirectional, adjustable percent
cutoff counting system. Basically, a bottom and top percentage will be
customizable by the user to determine cutoffs for the top and bottom
sizes of cells in the algorithm.